+++ megalinter/reporters/SarifReporter.py
@@ -214,7 +214,10 @@ class SarifReporter(Reporter):
                                     )
                                 result["locations"][id_location] = location
                         run["results"][id_result] = result
-
+                else:
+                    # make sure that there is a results entry so GitHub's SARIF validator doesn't cry
+                    run["results"] = []
+                    
                 # Update run in full list
                 linter_sarif_obj["runs"][id_run] = run
         return linter_sarif_obj